fix: stuff

+++ src/lexer/check.c
@@ -16,7 +16,7 @@ bool is_keyword(const char *start, size_t len)
 				    "unsigned", "void", "volatile", "while"
 				   };
 
-	for (int i = 0; i < sizeof(keywords) / sizeof(keywords[0]); i += 1) {
+	for (size_t i = 0; i < sizeof(keywords) / sizeof(keywords[0]); i += 1) {
 		if (strlen(keywords[i]) == len
 		    && strncmp(start, keywords[i], len) == 0)
 			return (true);
